# gfx2d: complete redesign of Sprite2D